Bug fixes associated with WRF-Chem/DART upgrade

+++ apm_run_scripts/RUN_REAL_TIME/FINAL_TEST_SCRIPTS/RS_Scripts/RS_TES_CO_Profile.ksh
@@ -102,55 +102,56 @@
       elif [[ -e ${TMP_OUTFILE} && -e ${OUTFILE} ]]; then
          cat ${OUTFILE} >> ${TMP_OUTFILE}
          rm -rf ${OUTFILE}
-      fi
+      fi 
       if [[ ! -e ${TMP_OUTFILE} ]]; then
          touch NO_TES_CO_${DATE}_DATA
       fi
 #
 # SET NAMELIST TO CONVERT TES CO ASCII TO OBS_SEQ 
-      export NL_FILEDIR=\'./\' 
-      export NL_FILENAME=\'${TMP_OUTFILE}\'
       export NL_FILEOUT=\'obs_seq_tes_co_profile_${DATE}.out\'
-      export NL_FAC_OBS_ERROR=${NL_FAC_OBS_ERROR_TES_CO}
-      export NL_USE_LOG_CO=${USE_LOG_CO_LOGIC}
-      export NL_USE_LOG_O3=${USE_LOG_O3_LOGIC}
-      export NL_USE_LOG_CO2=${USE_LOG_CO2_LOGIC}
-      export NL_USE_LOG_CH4=${USE_LOG_CH4_LOGIC}
-      export NL_USE_LOG_NH3=${USE_LOG_NH3_LOGIC}
+      if [[ -s ${TMP_OUTFILE} ]]; then
+         export NL_FILEDIR=\'./\' 
+         export NL_FILENAME=\'${TMP_OUTFILE}\'
+         export NL_FAC_OBS_ERROR=${NL_FAC_OBS_ERROR_TES_CO}
+         export NL_USE_LOG_CO=${USE_LOG_CO_LOGIC}
+         export NL_USE_LOG_O3=${USE_LOG_O3_LOGIC}
+         export NL_USE_LOG_CO2=${USE_LOG_CO2_LOGIC}
+         export NL_USE_LOG_CH4=${USE_LOG_CH4_LOGIC}
+         export NL_USE_LOG_NH3=${USE_LOG_NH3_LOGIC}
 #
 # MODEL PROFILE SETTINGS
-      export NL_PATH_MODEL=\'${WRFCHEM_TEMPLATE_DIR}\'
-      export NL_FILE_MODEL=\'${WRFCHEM_TEMPLATE_FILE}\'
-      export NL_NX_MODEL=${NNXP_CR}
-      export NL_NY_MODEL=${NNYP_CR}
-      export NL_NZ_MODEL=${NNZP_CR}
-#
-      export BIN_BEG_HH=${ASIM_MN_HH}
-      export BIN_BEG_MN=0
-      export BIN_BEG_SS=0
-      let HH_END=${ASIM_MX_HH}
-      let HHM_END=${HH_END}-1
-      export BIN_END_HH=${HHM_END}
-      export BIN_END_MN=59
-      export BIN_END_SS=59
-      let HH_BEG=${BIN_BEG_HH}
-      let MN_BEG=${BIN_BEG_MN}
-      let SS_BEG=${BIN_BEG_SS}
-      let HH_END=${BIN_END_HH}
-      let MN_END=${BIN_END_MN}
-      let SS_END=${BIN_END_SS}
-      let BIN_BEG_SEC=${HH_BEG}*3600+${MN_BEG}*60+${SS_BEG} 
-      let BIN_END_SEC=${HH_END}*3600+${MN_END}*60+${SS_END}
-      export NL_BIN_BEG_SEC=${BIN_BEG_SEC}
-      export NL_BIN_END_SEC=${BIN_END_SEC}
+         export NL_PATH_MODEL=\'${WRFCHEM_TEMPLATE_DIR}\'
+         export NL_FILE_MODEL=\'${WRFCHEM_TEMPLATE_FILE}\'
+         export NL_NX_MODEL=${NNXP_CR}
+         export NL_NY_MODEL=${NNYP_CR}
+         export NL_NZ_MODEL=${NNZP_CR}
+         export BIN_BEG_HH=${ASIM_MN_HH}
+         export BIN_BEG_MN=0
+         export BIN_BEG_SS=0
+         let HH_END=${ASIM_MX_HH}
+         let HHM_END=${HH_END}-1
+         export BIN_END_HH=${HHM_END}
+         export BIN_END_MN=59
+         export BIN_END_SS=59
+         let HH_BEG=${BIN_BEG_HH}
+         let MN_BEG=${BIN_BEG_MN}
+         let SS_BEG=${BIN_BEG_SS}
+         let HH_END=${BIN_END_HH}
+         let MN_END=${BIN_END_MN}
+         let SS_END=${BIN_END_SS}
+         let BIN_BEG_SEC=${HH_BEG}*3600+${MN_BEG}*60+${SS_BEG} 
+         let BIN_END_SEC=${HH_END}*3600+${MN_END}*60+${SS_END}
+         export NL_BIN_BEG_SEC=${BIN_BEG_SEC}
+         export NL_BIN_END_SEC=${BIN_END_SEC}
 #      
 # USE TES DATA 
-      rm -rf input.nml
-      ${NAMELIST_SCRIPTS_DIR}/OBS_CONVERTERS/da_create_dart_tes_input_nml.ksh
+         rm -rf input.nml
+         ${NAMELIST_SCRIPTS_DIR}/OBS_CONVERTERS/da_create_dart_tes_input_nml.ksh
 #
 # GET EXECUTABLE
-      cp ${DART_DIR}/observations/obs_converters/ATMOS_CHEM/TES_CO/work/tes_co_profile_ascii_to_obs ./.
-      ./tes_co_profile_ascii_to_obs > index.html 2>&1
+         cp ${DART_DIR}/observations/obs_converters/ATMOS_CHEM/TES_CO/work/tes_co_profile_ascii_to_obs ./.
+         ./tes_co_profile_ascii_to_obs > index.html 2>&1
+      fi
 #
 # COPY OUTPUT TO ARCHIVE LOCATION
       if [[ -s ${NL_FILEOUT} ]]; then
